refactor: adding typescript
