+++ next.config.ts
@@ -1,0 +1,6 @@
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {};
+
+export default nextConfig;
+
