+++ src/app/layout.tsx
@@ -1,7 +1,11 @@
 import "./globals.css";
+import { ReactNode } from "react";
 
-
-export default function RootLayout({ children }) {
+export default function RootLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
   return (
     <html lang="en">
       <body className={`antialiased`}>
@@ -9,4 +13,4 @@ export default function RootLayout({ children }) {
       </body>
     </html>
   );
-}
+}
