+++ src/components/ui/label.tsx
@@ -5,19 +5,25 @@ import * as LabelPrimitive from "@radix-ui/react-label"
 
 import { cn } from "@/lib/utils"
 
-function Label({
-  className,
-  ...props
-}) {
+interface LabelProps extends React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root> {}
+
+const Label = React.forwardRef<
+  React.ElementRef<typeof LabelPrimitive.Root>,
+  LabelProps
+>(({ className, ...props }, ref) => {
   return (
     <LabelPrimitive.Root
+      ref={ref}
       data-slot="label"
       className={cn(
         "flex items-center gap-2 text-sm leading-none font-medium select-none group-data-[disabled=true]:pointer-events-none group-data-[disabled=true]:opacity-50 peer-disabled:cursor-not-allowed peer-disabled:opacity-50",
         className
       )}
-      {...props} />
-  );
-}
+      {...props}
+    />
+  )
+})
+Label.displayName = "Label"
 
 export { Label }
+export type { LabelProps }
