+++ src/lib/utils.ts
@@ -1,0 +1,6 @@
+import { clsx, ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge"
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
